Fix a noisy error log about file deletion

+++ src/main/java/delta/games/lotro/tools/extraction/geo/markers/MainCategoriesPruner.java
@@ -57,10 +57,13 @@ public class MainCategoriesPruner
       {
         categoriesMgr.removeCategory(code);
         File iconFile=categoriesMgr.getIconFile(category);
-        boolean ok=iconFile.delete();
-        if (!ok)
+        if (iconFile.exists())
         {
-          LOGGER.warn("Failed to delete file: {}",iconFile);
+          boolean ok=iconFile.delete();
+          if (!ok)
+          {
+            LOGGER.warn("Failed to delete file: {}",iconFile);
+          }
         }
       }
     }
